+++ Lib/Math/math_common.h
@@ -20,7 +20,7 @@ namespace math {
     }
 
 
-    bool solveQuadratic(const float &a, const float &b, const float &c, float &x0, float &x1);
+    bool solveQuadratic(const double &a, const double &b, const double &c, double &x0, double &x1);
 
     // TODO All basic math function to remove cmath need
 }
